Update en/1.4/migration-guide/index.md
Fix requirements link

Note: This fixes it as a relative link on /1.4/migration-guide. This is
how the Migration Guide is linked from the 1.4 page and the sidebar
menu. However, the relative link would be invalid if the Migration
Guide is linked as /1.4/migration-guide/ (with the ending slash). I'm
not sure how to fix it for both.